:hammer: update logical_data
