+++ search/logical_data/new_logical_data.cpg
@@ -1,0 +1,1 @@
+UTF-8
